[Add] #4 - AppDelegate를 새로운 파일로 분리
