+++ Staccato-iOS/Staccato-iOS/App/Staccato_iOSApp.swift
@@ -5,8 +5,6 @@
 //  Created by Gyunni on 1/2/25.
 //
 
-import GoogleMaps
-
 import SwiftUI
 
 @main
@@ -17,18 +15,5 @@ struct Staccato_iOSApp: App {
         WindowGroup {
             ContentView()
         }
-    }
-}
-
-// Maps SDK 초기화를 위해 AppDelegate 구현
-class AppDelegate: NSObject, UIApplicationDelegate {
-    func application(
-        _ application: UIApplication,
-        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-    ) -> Bool {
-        let apiKey = Bundle.main.infoDictionary?["API_KEY"] as! String
-        GMSServices.provideAPIKey(apiKey)
-        
-        return true
     }
 }
